Crate icu_capi

This crate contains the source of truth for the Diplomat-generated FFI bindings. This generates the C, C++, JavaScript, and TypeScript bindings. This crate also contains the extern "C" FFI for ICU4X.

While the types in this crate are public, APIs from this crate are not intended to be used from Rust and as such this crate may unpredictably change its Rust API across compatible semver versions. The extern "C" APIs exposed by this crate, while not directly documented, are stable within the same major semver version, as are the bindings exposed under the cpp/ and js/ folders.

This crate may still be explored for documentation on docs.rs, and there are generated language-specific docs available as well. C++ has sphinx docs in cpp/docs/, and the header files also contain documentation comments. The JS version has sphinx docs under js/docs, and the TypeScript sources in js/include are compatible with tsdoc.

This crate is no_std and will not typically build as a staticlib on its own. If you wish to link to it you should prefer using icu_capi_staticlib, or for more esoteric platforms you may write a shim crate depending on this crate that hooks in an allocator and panic hook.

More information on using ICU4X from C++ can be found in our tutorial.

Modules

bidi

calendar

collator

data_struct

date

datetime

datetime_formatter

decimal

errors

fallbacker

fixed_decimal

list

locale

locid_transform

logging

normalizer

normalizer_properties

pluralrules

properties_maps

properties_sets

provider

script

segmenter_grapheme

segmenter_line

segmenter_sentence

segmenter_word

time

timezone

timezone_formatter

week

zoned_formatter