icu_capi::bidi

Module ffi

Structs

ICU4XBidi

An ICU4X Bidi object, containing loaded bidi data

ICU4XBidiInfo

An object containing bidi information for a given string, produced by for_text() on ICU4XBidi

ICU4XBidiParagraph

Bidi information for a single processed paragraph

ICU4XReorderedIndexMap

Thin wrapper around a vector that maps visual indices to source indices

Enums

ICU4XBidiDirection

Functions

ICU4XBidiDirection_destroy 🔒

ICU4XBidiInfo_destroy 🔒

ICU4XBidiInfo_level_at 🔒

ICU4XBidiInfo_paragraph_at 🔒

ICU4XBidiInfo_paragraph_count 🔒

ICU4XBidiInfo_size 🔒

ICU4XBidiParagraph_destroy 🔒

ICU4XBidiParagraph_direction 🔒

ICU4XBidiParagraph_level_at 🔒

ICU4XBidiParagraph_range_end 🔒

ICU4XBidiParagraph_range_start 🔒

ICU4XBidiParagraph_reorder_line 🔒

ICU4XBidiParagraph_set_paragraph_in_text 🔒

ICU4XBidiParagraph_size 🔒

ICU4XBidi_create 🔒

ICU4XBidi_destroy 🔒

ICU4XBidi_for_text 🔒

ICU4XBidi_for_text_valid_utf8 🔒

ICU4XBidi_level_is_ltr 🔒

ICU4XBidi_level_is_rtl 🔒

ICU4XBidi_level_ltr 🔒

ICU4XBidi_level_rtl 🔒

ICU4XBidi_reorder_visual 🔒

ICU4XReorderedIndexMap_as_slice 🔒

ICU4XReorderedIndexMap_destroy 🔒

ICU4XReorderedIndexMap_get 🔒

ICU4XReorderedIndexMap_is_empty 🔒

ICU4XReorderedIndexMap_len 🔒