icu_capi::bidi::ffi

Struct ICU4XBidiParagraph

pub struct ICU4XBidiParagraph<'info>(pub Paragraph<'info, 'info>);

Bidi information for a single processed paragraph

Tuple Fields
0: Paragraph<'info, 'info>
Implementations

impl<'info> ICU4XBidiParagraph<'info>

pub fn set_paragraph_in_text(&mut self, n: usize) -> Result<(), ICU4XError>

Given a paragraph index n within the surrounding text, this sets this object to the paragraph at that index. Returns ICU4XError::OutOfBoundsError when out of bounds.

This is equivalent to calling paragraph_at() on ICU4XBidiInfo but doesn’t create a new object

pub fn direction(&self) -> ICU4XBidiDirection

The primary direction of this paragraph

pub fn size(&self) -> usize

The number of bytes in this paragraph

pub fn range_start(&self) -> usize

The start index of this paragraph within the source text

pub fn range_end(&self) -> usize

The end index of this paragraph within the source text

pub fn reorder_line( &self, range_start: usize, range_end: usize, out: &mut DiplomatWriteable, ) -> Result<(), ICU4XError>

Reorder a line based on display order. The ranges are specified relative to the source text and must be contained within this paragraph’s range.

pub fn level_at(&self, pos: usize) -> u8

Get the BIDI level at a particular byte index in this paragraph. This integer is conceptually a unicode_bidi::Level, and can be further inspected using the static methods on ICU4XBidi.

Returns 0 (equivalent to Level::ltr()) on error