Enum icu_capi::calendar::ffi::ICU4XAnyCalendarKind

#[repr(C)]
pub enum ICU4XAnyCalendarKind {

    Iso = 0,
    Gregorian = 1,
    Buddhist = 2,
    Japanese = 3,
    JapaneseExtended = 4,
    Ethiopian = 5,
    EthiopianAmeteAlem = 6,
    Indian = 7,
    Coptic = 8,
    Dangi = 9,
    Chinese = 10,
    Hebrew = 11,
    IslamicCivil = 12,
    IslamicObservational = 13,
    IslamicTabular = 14,
    IslamicUmmAlQura = 15,
    Persian = 16,
    Roc = 17,

}

The various calendar types currently supported by ICU4XCalendar

Variants

Iso = 0

The kind of an Iso calendar

Gregorian = 1

The kind of a Gregorian calendar

Buddhist = 2

The kind of a Buddhist calendar

Japanese = 3

The kind of a Japanese calendar with modern eras

JapaneseExtended = 4

The kind of a Japanese calendar with modern and historic eras

Ethiopian = 5

The kind of an Ethiopian calendar, with Amete Mihret era

EthiopianAmeteAlem = 6

The kind of an Ethiopian calendar, with Amete Alem era

Indian = 7

The kind of a Indian calendar

Coptic = 8

The kind of a Coptic calendar

Dangi = 9

The kind of a Dangi calendar

Chinese = 10

The kind of a Chinese calendar

Hebrew = 11

The kind of a Hebrew calendar

IslamicCivil = 12

The kind of a Islamic civil calendar

IslamicObservational = 13

The kind of a Islamic observational calendar

IslamicTabular = 14

The kind of a Islamic tabular calendar

IslamicUmmAlQura = 15

The kind of a Islamic Umm al-Qura calendar

Persian = 16

The kind of a Persian calendar

Roc = 17

The kind of a Roc calendar

Implementations

impl ICU4XAnyCalendarKind

pub fn get_for_locale(locale: &ICU4XLocale) -> Option<ICU4XAnyCalendarKind>

Read the calendar type off of the -u-ca- extension on a locale.

Returns nothing if there is no calendar on the locale or if the locale’s calendar is not known or supported.

pub fn get_for_bcp47(s: &DiplomatStr) -> Option<ICU4XAnyCalendarKind>

Obtain the calendar type given a BCP-47 -u-ca- extension string.

Returns nothing if the calendar is not known or supported.

pub fn bcp47(self, write: &mut DiplomatWriteable) -> Result<(), ICU4XError>

Obtain the string suitable for use in the -u-ca- extension in a BCP47 locale.