Struct icu_capi::collections_sets::ffi::ICU4XCodePointSetBuilder

pub struct ICU4XCodePointSetBuilder(pub CodePointInversionListBuilder);

Tuple Fields
0: CodePointInversionListBuilder
Implementations

impl ICU4XCodePointSetBuilder

pub fn create() -> Box<Self>

Make a new set builder containing nothing

pub fn build(&mut self) -> Box<ICU4XCodePointSetData>

Build this into a set

This object is repopulated with an empty builder

pub fn complement(&mut self)

Complements this set

(Elements in this set are removed and vice versa)

pub fn is_empty(&self) -> bool

Returns whether this set is empty

pub fn add_char(&mut self, ch: DiplomatChar)

Add a single character to the set

pub fn add_u32(&mut self, ch: u32)

Deprecated, use add_char.

pub fn add_inclusive_range(&mut self, start: DiplomatChar, end: DiplomatChar)

Add an inclusive range of characters to the set

pub fn add_inclusive_range_u32(&mut self, start: u32, end: u32)

Deprecated, use add_inclusive_range.

pub fn add_set(&mut self, data: &ICU4XCodePointSetData)

Add all elements that belong to the provided set to the set

pub fn remove_char(&mut self, ch: DiplomatChar)

Remove a single character to the set

pub fn remove_inclusive_range(&mut self, start: DiplomatChar, end: DiplomatChar)

Remove an inclusive range of characters from the set

pub fn remove_set(&mut self, data: &ICU4XCodePointSetData)

Remove all elements that belong to the provided set from the set

pub fn retain_char(&mut self, ch: DiplomatChar)

Removes all elements from the set except a single character

pub fn retain_inclusive_range(&mut self, start: DiplomatChar, end: DiplomatChar)

Removes all elements from the set except an inclusive range of characters f

pub fn retain_set(&mut self, data: &ICU4XCodePointSetData)

Removes all elements from the set except all elements in the provided set

pub fn complement_char(&mut self, ch: DiplomatChar)

Complement a single character to the set

(Characters which are in this set are removed and vice versa)

pub fn complement_inclusive_range( &mut self, start: DiplomatChar, end: DiplomatChar, )

Complement an inclusive range of characters from the set

(Characters which are in this set are removed and vice versa)

pub fn complement_set(&mut self, data: &ICU4XCodePointSetData)

Complement all elements that belong to the provided set from the set

(Characters which are in this set are removed and vice versa)