Module icu_capi::custom_writeable

Structs

ICU4XWriteable
An object that can one can write UTF-8 strings to

Functions

icu4x_buffer_writeable_create
Create an ICU4XWriteable that can write to a dynamically allocated buffer managed by Rust.

icu4x_buffer_writeable_destroy⚠
Destructor for Rust-memory backed writables.

icu4x_buffer_writeable_get_bytes
Grabs a pointer to the underlying buffer of a writable.

icu4x_buffer_writeable_len
Gets the length in bytes of the content written to the writable.

icu4x_simple_writeable⚠
Create an ICU4XWriteable that can write to a fixed-length stack allocated u8 buffer.