Function icu_capi::custom_writeable::icu4x_buffer_writeable_len

#[no_mangle]pub extern "C" fn icu4x_buffer_writeable_len(this: &ICU4XWriteable) -> usize

Gets the length in bytes of the content written to the writable.

Safety

this must be a pointer to a valid ICU4XWriteable constructed by icu4x_buffer_writeable_create().