Struct icu_capi::date::ffi::ICU4XDate

pub struct ICU4XDate(pub Date<Arc<AnyCalendar>>);

An ICU4X Date object capable of containing a date and time for any calendar.

Tuple Fields
0: Date<Arc<AnyCalendar>>
Implementations

impl<'transparent_convert_outer> ICU4XDate

pub(crate) fn transparent_convert( from: &'transparent_convert_outer Date<Arc<AnyCalendar>>, ) -> &'transparent_convert_outer Self

impl ICU4XDate

pub fn create_from_iso_in_calendar( year: i32, month: u8, day: u8, calendar: &ICU4XCalendar, ) -> Result<Box<ICU4XDate>, ICU4XError>

Creates a new ICU4XDate representing the ISO date and time given but in a given calendar

pub fn create_from_codes_in_calendar( era_code: &DiplomatStr, year: i32, month_code: &DiplomatStr, day: u8, calendar: &ICU4XCalendar, ) -> Result<Box<ICU4XDate>, ICU4XError>

Creates a new ICU4XDate from the given codes, which are interpreted in the given calendar system

pub fn to_calendar(&self, calendar: &ICU4XCalendar) -> Box<ICU4XDate>

Convert this date to one in a different calendar

pub fn to_iso(&self) -> Box<ICU4XIsoDate>

Converts this date to ISO

pub fn day_of_year(&self) -> u16

Returns the 1-indexed day in the year for this date

pub fn day_of_month(&self) -> u32

Returns the 1-indexed day in the month for this date

pub fn day_of_week(&self) -> ICU4XIsoWeekday

Returns the day in the week for this day

pub fn week_of_month(&self, first_weekday: ICU4XIsoWeekday) -> u32

Returns the week number in this month, 1-indexed, based on what is considered the first day of the week (often a locale preference).

first_weekday can be obtained via first_weekday() on ICU4XWeekCalculator

pub fn week_of_year( &self, calculator: &ICU4XWeekCalculator, ) -> Result<ICU4XWeekOf, ICU4XError>

Returns the week number in this year, using week data

pub fn ordinal_month(&self) -> u32

Returns 1-indexed number of the month of this date in its year

Note that for lunar calendars this may not lead to the same month having the same ordinal month across years; use month_code if you care about month identity.

pub fn month_code( &self, write: &mut DiplomatWriteable, ) -> Result<(), ICU4XError>

Returns the month code for this date. Typically something like “M01”, “M02”, but can be more complicated for lunar calendars.

pub fn year_in_era(&self) -> i32

Returns the year number in the current era for this date

pub fn era(&self, write: &mut DiplomatWriteable) -> Result<(), ICU4XError>

Returns the era for this date,

pub fn months_in_year(&self) -> u8

Returns the number of months in the year represented by this date

pub fn days_in_month(&self) -> u8

Returns the number of days in the month represented by this date

pub fn days_in_year(&self) -> u16

Returns the number of days in the year represented by this date

pub fn calendar(&self) -> Box<ICU4XCalendar>

Returns the ICU4XCalendar object backing this date