Struct icu_capi::datetime::ffi::ICU4XDateTime

pub struct ICU4XDateTime(pub DateTime<Arc<AnyCalendar>>);

An ICU4X DateTime object capable of containing a date and time for any calendar.

Tuple Fields
0: DateTime<Arc<AnyCalendar>>
Implementations

impl ICU4XDateTime

pub fn create_from_iso_in_calendar(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8,
nanosecond: u32,
calendar: &ICU4XCalendar
) -> DiplomatResult<Box<ICU4XDateTime>, ICU4XError>

Creates a new ICU4XDateTime representing the ISO date and time given but in a given calendar

pub fn create_from_codes_in_calendar(
era_code: &str,
year: i32,
month_code: &str,
day: u8,
hour: u8,
minute: u8,
second: u8,
nanosecond: u32,
calendar: &ICU4XCalendar
) -> DiplomatResult<Box<ICU4XDateTime>, ICU4XError>

Creates a new ICU4XDateTime from the given codes, which are interpreted in the given calendar system

pub fn create_from_date_and_time(
date: &ICU4XDate,
time: &ICU4XTime
) -> Box<ICU4XDateTime>

Creates a new ICU4XDateTime from an ICU4XDate and ICU4XTime object

pub fn date(&self) -> Box<ICU4XDate>

Gets a copy of the date contained in this object

pub fn time(&self) -> Box<ICU4XTime>

Gets the time contained in this object

pub fn to_iso(&self) -> Box<ICU4XIsoDateTime>

Converts this date to ISO

pub fn to_calendar(&self, calendar: &ICU4XCalendar) -> Box<ICU4XDateTime>

Convert this datetime to one in a different calendar

pub fn hour(&self) -> u8

Returns the hour in this time

pub fn minute(&self) -> u8

Returns the minute in this time

pub fn second(&self) -> u8

Returns the second in this time

pub fn nanosecond(&self) -> u32

Returns the nanosecond in this time

pub fn day_of_month(&self) -> u32

Returns the 1-indexed day in the month for this date

pub fn day_of_week(&self) -> ICU4XIsoWeekday

Returns the day in the week for this day

pub fn week_of_month(&self, first_weekday: ICU4XIsoWeekday) -> u32

Returns the week number in this month, 1-indexed, based on what is considered the first day of the week (often a locale preference).

first_weekday can be obtained via first_weekday() on ICU4XWeekCalculator

pub fn week_of_year(
&self,
calculator: &ICU4XWeekCalculator
) -> DiplomatResult<ICU4XWeekOf, ICU4XError>

Returns the week number in this year, using week data

pub fn ordinal_month(&self) -> u32

Returns 1-indexed number of the month of this date in its year

Note that for lunar calendars this may not lead to the same month having the same ordinal month across years; use month_code if you care about month identity.

pub fn month_code(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XError>

Returns the month code for this date. Typically something like “M01”, “M02”, but can be more complicated for lunar calendars.

pub fn year_in_era(&self) -> i32

Returns the year number in the current era for this date

pub fn era(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XError>

Returns the era for this date,

pub fn months_in_year(&self) -> u8

Returns the number of months in the year represented by this date

pub fn days_in_month(&self) -> u8

Returns the number of days in the month represented by this date

pub fn days_in_year(&self) -> u32

Returns the number of days in the year represented by this date

pub fn calendar(&self) -> Box<ICU4XCalendar>

Returns the ICU4XCalendar object backing this date