Struct icu_capi::datetime::ffi::ICU4XIsoDateTime

pub struct ICU4XIsoDateTime(pub DateTime<Iso>);

An ICU4X DateTime object capable of containing a ISO-8601 date and time.

Tuple Fields
0: DateTime<Iso>
Implementations

impl ICU4XIsoDateTime

pub fn create(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8,
nanosecond: u32
) -> DiplomatResult<Box<ICU4XIsoDateTime>, ICU4XError>

Creates a new ICU4XIsoDateTime from the specified date and time.

pub fn crate_from_date_and_time(
date: &ICU4XIsoDate,
time: &ICU4XTime
) -> Box<ICU4XIsoDateTime>

Creates a new ICU4XIsoDateTime from an ICU4XIsoDate and ICU4XTime object

pub fn create_from_minutes_since_local_unix_epoch(
minutes: i32
) -> DiplomatResult<Box<ICU4XIsoDateTime>, ICU4XError>

Construct from the minutes since the local unix epoch for this date (Jan 1 1970, 00:00)

pub fn date(&self) -> Box<ICU4XIsoDate>

Gets the date contained in this object

pub fn time(&self) -> Box<ICU4XTime>

Gets the time contained in this object

pub fn to_any(&self) -> Box<ICU4XDateTime>

Converts this to an ICU4XDateTime capable of being mixed with dates of other calendars

pub fn minutes_since_local_unix_epoch(&self) -> i32

Gets the minutes since the local unix epoch for this date (Jan 1 1970, 00:00)

pub fn to_calendar(&self, calendar: &ICU4XCalendar) -> Box<ICU4XDateTime>

Convert this datetime to one in a different calendar

pub fn hour(&self) -> u8

Returns the hour in this time

pub fn minute(&self) -> u8

Returns the minute in this time

pub fn second(&self) -> u8

Returns the second in this time

pub fn nanosecond(&self) -> u32

Returns the nanosecond in this time

pub fn day_of_month(&self) -> u32

Returns the 1-indexed day in the month for this date

pub fn day_of_week(&self) -> ICU4XIsoWeekday

Returns the day in the week for this day

pub fn week_of_month(&self, first_weekday: ICU4XIsoWeekday) -> u32

Returns the week number in this month, 1-indexed, based on what is considered the first day of the week (often a locale preference).

first_weekday can be obtained via first_weekday() on ICU4XWeekCalculator

pub fn week_of_year(
&self,
calculator: &ICU4XWeekCalculator
) -> DiplomatResult<ICU4XWeekOf, ICU4XError>

Returns the week number in this year, using week data

pub fn month(&self) -> u32

Returns 1-indexed number of the month of this date in its year

pub fn year(&self) -> i32

Returns the year number for this date

pub fn months_in_year(&self) -> u8

Returns the number of months in the year represented by this date

pub fn days_in_month(&self) -> u8

Returns the number of days in the month represented by this date

pub fn days_in_year(&self) -> u32

Returns the number of days in the year represented by this date