Enum icu_capi::errors::ffi::ICU4XError

#[repr(C)]
#[repr(C)]
pub enum ICU4XError {

    UnknownError,
    WriteableError,
    OutOfBoundsError,
    DataMissingResourceKeyError,
    DataMissingVariantError,
    DataMissingLocaleError,
    DataMissingResourceOptionsError,
    DataNeedsVariantError,
    DataNeedsLocaleError,
    DataExtraneousResourceOptionsError,
    DataFilteredResourceError,
    DataMismatchedTypeError,
    DataMissingPayloadError,
    DataInvalidStateError,
    DataCustomError,
    DataIoError,
    DataUnavailableBufferFormatError,
    LocaleUndefinedSubtagError,
    LocaleParserError,
    DataStructValidityError,
    PropertyUnknownScriptIdError,
    PropertyUnknownGeneralCategoryGroupError,
    DecimalLimitError,
    DecimalSyntaxError,
    PluralParserError,
    DateTimeParseError,
    DateTimeOverflowError,
    DateTimeUnderflowError,
    DateTimeInvalidTimeZoneOffsetError,
    DateTimeOutOfRangeError,
    DateTimeMissingInputError,
    DateTimeFormatPatternError,
    DateTimeFormatMissingInputFieldError,
    DateTimeFormatSkeletonError,
    DateTimeFormatUnsupportedFieldError,
    DateTimeFormatUnsupportedOptionsError,
    DateTimeFormatMissingWeekdaySymbolError,
    DateTimeFormatMissingMonthSymbolError,
    DateTimeFormatMissingEraSymbolError,
    DateTimeFormatFixedDecimalError,
    DateTimeFormatMismatchedAnyCalendarError,

}

A common enum for errors that ICU4X may return, organized by API

The error names are stable and can be checked against as strings in the JS API

Variants

UnknownError

The error is not currently categorized as ICU4XError. Please file a bug

WriteableError

An error arising from writing to a string Typically found when not enough space is allocated Most APIs that return a string may return this error

OutOfBoundsError

DataMissingResourceKeyError

DataMissingVariantError

DataMissingLocaleError

DataMissingResourceOptionsError

DataNeedsVariantError

DataNeedsLocaleError

DataExtraneousResourceOptionsError

DataFilteredResourceError

DataMismatchedTypeError

DataMissingPayloadError

DataInvalidStateError

DataCustomError

DataIoError

DataUnavailableBufferFormatError

LocaleUndefinedSubtagError

The subtag being requested was not set

LocaleParserError

The locale or subtag string failed to parse

DataStructValidityError

Attempted to construct an invalid data struct

PropertyUnknownScriptIdError

PropertyUnknownGeneralCategoryGroupError

DecimalLimitError

DecimalSyntaxError

PluralParserError

DateTimeParseError

DateTimeOverflowError

DateTimeUnderflowError

DateTimeInvalidTimeZoneOffsetError

DateTimeOutOfRangeError

DateTimeMissingInputError

DateTimeFormatPatternError

DateTimeFormatMissingInputFieldError

DateTimeFormatSkeletonError

DateTimeFormatUnsupportedFieldError

DateTimeFormatUnsupportedOptionsError

DateTimeFormatMissingWeekdaySymbolError

DateTimeFormatMissingMonthSymbolError

DateTimeFormatMissingEraSymbolError

DateTimeFormatFixedDecimalError

DateTimeFormatMismatchedAnyCalendarError