Struct icu_capi::fixed_decimal::ffi::ICU4XFixedDecimal

#[repr(transparent)]
pub struct ICU4XFixedDecimal(pub FixedDecimal);

A decimal number. See the Rust docs for more information.

Tuple Fields
0: FixedDecimal
Implementations

impl ICU4XFixedDecimal

pub fn create(v: i32) -> Box<ICU4XFixedDecimal>

Construct an ICU4XFixedDecimal from an integer.

See the Rust docs for more information.

pub fn create_from_f64_with_max_precision(
f: f64
) -> Option<Box<ICU4XFixedDecimal>>

Construct an ICU4XFixedDecimal from an float, with enough digits to recover the original floating point in IEEE 754 without needing trailing zeros

See the Rust docs for more information.

pub fn create_from_f64_with_lower_magnitude(
f: f64,
precision: i16,
rounding_mode: ICU4XFixedDecimalRoundingMode
) -> Option<Box<ICU4XFixedDecimal>>

Construct an ICU4XFixedDecimal from an float, with a given power of 10 for the lower magnitude

See the Rust docs for more information.

pub fn create_from_f64_with_significant_digits(
f: f64,
digits: u8,
rounding_mode: ICU4XFixedDecimalRoundingMode
) -> Option<Box<ICU4XFixedDecimal>>

Construct an ICU4XFixedDecimal from an float, for a given number of significant digits

See the Rust docs for more information.

pub fn create_fromstr(v: &str) -> ICU4XCreateFixedDecimalResult

Construct an ICU4XFixedDecimal from a string. See the Rust docs for more information.

pub fn multiply_pow10(&mut self, power: i16) -> bool

Multiply the ICU4XFixedDecimal by a given power of ten. See the Rust docs for more information.

pub fn negate(&mut self)

Invert the sign of the ICU4XFixedDecimal. See the Rust docs for more information.

pub fn pad_left(&mut self, digits: u16)

Zero-pad the ICU4XFixedDecimal on the left to a particular number of integer digits See the Rust docs for more information.

pub fn truncate_left(&mut self, magnitude: i16)

Truncate the ICU4XFixedDecimal on the left to a particular magnitude, deleting digits if necessary. This is useful for, e.g. abbreviating years (“2022” -> “22”) See the Rust docs for more information.

pub fn pad_right(&mut self, negative_magnitude: u16)

Zero-pad the ICU4XFixedDecimal on the right to a particular (negative) magnitude See the Rust docs for more information.

pub fn to_string(&self, to: &mut DiplomatWriteable)

Format the ICU4XFixedDecimal as a string. See the Rust docs for more information.