Struct icu_capi::locale::ffi::ICU4XLocale

#[repr(transparent)]
pub struct ICU4XLocale(pub Locale);

An ICU4X Locale, capable of representing strings like "en-US". See the Rust docs for more information.

Implementations

impl ICU4XLocale

pub fn create(name: &str) -> Option<Box<ICU4XLocale>>

Construct an ICU4XLocale from an locale identifier. See the Rust docs for more information.

pub fn create_en() -> Box<ICU4XLocale>

Construct an ICU4XLocale for the English language.

pub fn create_bn() -> Box<ICU4XLocale>

Construct an ICU4XLocale for the Bangla language.

pub fn clone(&self) -> Box<ICU4XLocale>

Clones the ICU4XLocale. See the Rust docs for more information.

pub fn basename(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XLocaleError>

Write a string representation of the LanguageIdentifier part of ICU4XLocale to write. See the Rust docs for more information.

pub fn get_unicode_extension(
&self,
bytes: &str,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XLocaleError>

Write a string representation of the unicode extension to write See the Rust docs for more information.

pub fn language(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XLocaleError>

Write a string representation of ICU4XLocale language to write See the Rust docs for more information.

pub fn region(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XLocaleError>

Write a string representation of ICU4XLocale region to write See the Rust docs for more information.

pub fn script(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XLocaleError>

Write a string representation of ICU4XLocale script to write See the Rust docs for more information.

pub fn tostring(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ICU4XLocaleError>

Write a string representation of ICU4XLocale to write See the Rust docs for more information.