Module icu_capi::normalizer_properties::ffi

Structs

ICU4XCanonicalCombiningClassMap

Lookup of the Canonical_Combining_Class Unicode property

ICU4XCanonicalComposition

The raw canonical composition operation.

ICU4XCanonicalDecomposition

The raw (non-recursive) canonical decomposition operation.

ICU4XDecomposed

The outcome of non-recursive canonical decomposition of a character. second will be NUL when the decomposition expands to a single character (which may or may not be the original one)

Functions

ICU4XCanonicalCombiningClassMap_create 🔒

ICU4XCanonicalCombiningClassMap_destroy 🔒

ICU4XCanonicalCombiningClassMap_get 🔒

ICU4XCanonicalCombiningClassMap_get32 🔒

ICU4XCanonicalComposition_compose 🔒

ICU4XCanonicalComposition_create 🔒

ICU4XCanonicalComposition_destroy 🔒

ICU4XCanonicalDecomposition_create 🔒

ICU4XCanonicalDecomposition_decompose 🔒

ICU4XCanonicalDecomposition_destroy 🔒

ICU4XDecomposed_destroy 🔒