Function icu_capi::pluralrules::icu4x_plural_rules_create

#[no_mangle]pub extern "C" fn icu4x_plural_rules_create(
    locale: &ICULocale, 
    provider: &ICU4XDataProvider, 
    ty: ICU4XPluralRuleType
) -> ICU4XCreatePluralRulesResult

FFI version of PluralRules::try_new() see its docs for more details

Safety

locale should be constructed via icu4x_locale_create()

provider should be constructed via one of the functions in crate::locale

Only access rules in the result if success is true.