icu_capi::properties_maps

Module ffi

Structs

ICU4XCodePointMapData8

An ICU4X Unicode Map Property object, capable of querying whether a code point (key) to obtain the Unicode property value, for a specific Unicode property.

ICU4XCodePointMapData16

An ICU4X Unicode Map Property object, capable of querying whether a code point (key) to obtain the Unicode property value, for a specific Unicode property.

Functions

ICU4XCodePointMapData8_destroy 🔒

ICU4XCodePointMapData8_general_category_to_mask 🔒

ICU4XCodePointMapData8_get 🔒

ICU4XCodePointMapData8_get32 🔒

ICU4XCodePointMapData8_get_set_for_value 🔒

ICU4XCodePointMapData8_iter_ranges_for_mask 🔒

ICU4XCodePointMapData8_iter_ranges_for_value 🔒

ICU4XCodePointMapData8_iter_ranges_for_value_complemented 🔒

ICU4XCodePointMapData8_load_bidi_class 🔒

ICU4XCodePointMapData8_load_east_asian_width 🔒

ICU4XCodePointMapData8_load_general_category 🔒

ICU4XCodePointMapData8_load_hangul_syllable_type 🔒

ICU4XCodePointMapData8_load_indic_syllabic_category 🔒

ICU4XCodePointMapData8_load_joining_type 🔒

ICU4XCodePointMapData8_load_line_break 🔒

ICU4XCodePointMapData8_load_sentence_break 🔒

ICU4XCodePointMapData8_load_word_break 🔒

ICU4XCodePointMapData8_try_grapheme_cluster_break 🔒

ICU4XCodePointMapData16_destroy 🔒

ICU4XCodePointMapData16_get 🔒

ICU4XCodePointMapData16_get32 🔒

ICU4XCodePointMapData16_get_set_for_value 🔒

ICU4XCodePointMapData16_iter_ranges_for_value 🔒

ICU4XCodePointMapData16_iter_ranges_for_value_complemented 🔒

ICU4XCodePointMapData16_load_script 🔒

convert_8 🔒