Module icu_capi::properties_names::ffi

Structs

ICU4XGeneralCategoryNameToMaskMapper

A type capable of looking up General Category mask values from a string name.

ICU4XPropertyValueNameToEnumMapper

A type capable of looking up a property value from a string name.

Functions

ICU4XGeneralCategoryNameToMaskMapper_destroy 🔒

ICU4XGeneralCategoryNameToMaskMapper_get_loose 🔒

ICU4XGeneralCategoryNameToMaskMapper_get_strict 🔒

ICU4XGeneralCategoryNameToMaskMapper_load 🔒

ICU4XPropertyValueNameToEnumMapper_destroy 🔒

ICU4XPropertyValueNameToEnumMapper_get_loose 🔒

ICU4XPropertyValueNameToEnumMapper_get_strict 🔒

ICU4XPropertyValueNameToEnumMapper_load_bidi_class 🔒

ICU4XPropertyValueNameToEnumMapper_load_east_asian_width 🔒

ICU4XPropertyValueNameToEnumMapper_load_general_category 🔒

ICU4XPropertyValueNameToEnumMapper_load_grapheme_cluster_break 🔒

ICU4XPropertyValueNameToEnumMapper_load_hangul_syllable_type 🔒

ICU4XPropertyValueNameToEnumMapper_load_indic_syllabic_category 🔒

ICU4XPropertyValueNameToEnumMapper_load_line_break 🔒

ICU4XPropertyValueNameToEnumMapper_load_script 🔒

ICU4XPropertyValueNameToEnumMapper_load_sentence_break 🔒

ICU4XPropertyValueNameToEnumMapper_load_word_break 🔒