Struct icu_capi::properties_sets::ffi::ICU4XCodePointSetData

pub struct ICU4XCodePointSetData(pub CodePointSetData);

An ICU4X Unicode Set Property object, capable of querying whether a code point is contained in a set based on a Unicode property.

Tuple Fields
0: CodePointSetData
Implementations

impl ICU4XCodePointSetData

pub fn contains(&self, cp: char) -> bool

Checks whether the code point is in the set.

pub fn contains32(&self, cp: u32) -> bool

Checks whether the code point (specified as a 32 bit integer, in UTF-32) is in the set.

pub fn load_for_general_category_group(
provider: &ICU4XDataProvider,
group: u32
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

which is a mask with the same format as the U_GC_XX_MASK mask in ICU4C

pub fn load_ascii_hex_digit(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_alnum(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_alphabetic(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_bidi_control(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_bidi_mirrored(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_blank(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_cased(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_case_ignorable(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_full_composition_exclusion(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_changes_when_casefolded(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_changes_when_casemapped(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_changes_when_nfkc_casefolded(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_changes_when_lowercased(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_changes_when_titlecased(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_changes_when_uppercased(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_dash(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_deprecated(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_default_ignorable_code_point(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_diacritic(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_emoji_modifier_base(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_emoji_component(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_emoji_modifier(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_emoji(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_emoji_presentation(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_extender(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_extended_pictographic(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_graph(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_grapheme_base(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_grapheme_extend(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_grapheme_link(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_hex_digit(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_hyphen(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_id_continue(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_ideographic(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_id_start(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_ids_binary_operator(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_ids_trinary_operator(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_join_control(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_logical_order_exception(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_lowercase(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_math(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_noncharacter_code_point(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_nfc_inert(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_nfd_inert(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_nfkc_inert(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_nfkd_inert(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_pattern_syntax(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_pattern_white_space(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_prepended_concatenation_mark(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_print(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_quotation_mark(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_radical(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_regional_indicator(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_soft_dotted(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_segment_starter(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_case_sensitive(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_sentence_terminal(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_terminal_punctuation(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_unified_ideograph(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_uppercase(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_variation_selector(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_white_space(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_xdigit(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_xid_continue(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>

pub fn load_xid_start(
provider: &ICU4XDataProvider
) -> DiplomatResult<Box<ICU4XCodePointSetData>, ICU4XError>