icu_capi::properties_unisets::ffi

Struct ICU4XUnicodeSetData

pub struct ICU4XUnicodeSetData(pub UnicodeSetData);

An ICU4X Unicode Set Property object, capable of querying whether a code point is contained in a set based on a Unicode property.

Tuple Fields
0: UnicodeSetData
Implementations

impl ICU4XUnicodeSetData

pub fn contains(&self, s: &DiplomatStr) -> bool

Checks whether the string is in the set.

pub fn contains_char(&self, cp: DiplomatChar) -> bool

Checks whether the code point is in the set.

pub fn contains32(&self, cp: u32) -> bool

Checks whether the code point (specified as a 32 bit integer, in UTF-32) is in the set.

pub fn load_basic_emoji( provider: &ICU4XDataProvider, ) -> Result<Box<ICU4XUnicodeSetData>, ICU4XError>

pub fn load_exemplars_main( provider: &ICU4XDataProvider, locale: &ICU4XLocale, ) -> Result<Box<ICU4XUnicodeSetData>, ICU4XError>

pub fn load_exemplars_auxiliary( provider: &ICU4XDataProvider, locale: &ICU4XLocale, ) -> Result<Box<ICU4XUnicodeSetData>, ICU4XError>

pub fn load_exemplars_punctuation( provider: &ICU4XDataProvider, locale: &ICU4XLocale, ) -> Result<Box<ICU4XUnicodeSetData>, ICU4XError>

pub fn load_exemplars_numbers( provider: &ICU4XDataProvider, locale: &ICU4XLocale, ) -> Result<Box<ICU4XUnicodeSetData>, ICU4XError>

pub fn load_exemplars_index( provider: &ICU4XDataProvider, locale: &ICU4XLocale, ) -> Result<Box<ICU4XUnicodeSetData>, ICU4XError>