Function icu_capi::provider::icu4x_fs_data_provider_create

#[no_mangle]pub unsafe extern "C" fn icu4x_fs_data_provider_create(
    path: *const u8, 
    len: usize
) -> ICU4XCreateDataProviderResult

Constructs an [FsDataProvider] and retirns it as an ICU4XDataProvider. See [FsDataProvider::try_new()] for more details.

Safety

path and len must point to a valid UTF-8 string, with len not including a null terminator if any.

Only access provider in the result if success is true.