Struct icu_capi::timezone::ffi::ICU4XCustomTimeZone

pub struct ICU4XCustomTimeZone(pub CustomTimeZone);

Tuple Fields
0: CustomTimeZone
Implementations

impl ICU4XCustomTimeZone

pub fn create_from_str(
s: &str
) -> DiplomatResult<Box<ICU4XCustomTimeZone>, ICU4XError>

pub fn try_set_gmt_offset_seconds(
&mut self,
offset_seconds: i32
) -> DiplomatResult<(), ICU4XError>

Sets the gmt_offset field from offset seconds.

Errors if the offset seconds are out of range.

pub fn clear_gmt_offset(&mut self)

Clears the gmt_offset field.

pub fn gmt_offset_seconds(&self) -> DiplomatResult<i32, ()>

Returns the value of the gmt_offset field as offset seconds.

Errors if the gmt_offset field is empty.

pub fn is_gmt_offset_positive(&self) -> DiplomatResult<bool, ()>

Returns whether the gmt_offset field is positive.

Errors if the gmt_offset field is empty.

pub fn is_gmt_offset_zero(&self) -> DiplomatResult<bool, ()>

Returns whether the gmt_offset field is zero.

Errors if the gmt_offset field is empty (which is not the same as zero).

pub fn gmt_offset_has_minutes(&self) -> DiplomatResult<bool, ()>

Returns whether the gmt_offset field has nonzero minutes.

Errors if the gmt_offset field is empty.

pub fn gmt_offset_has_seconds(&self) -> DiplomatResult<bool, ()>

Returns whether the gmt_offset field has nonzero seconds.

Errors if the gmt_offset field is empty.

pub fn try_set_time_zone_id(
&mut self,
id: &str
) -> DiplomatResult<(), ICU4XError>

Sets the time_zone_id field from a string.

Errors if the string is not a valid BCP-47 time zone ID.

pub fn clear_time_zone_id(&mut self)

Clears the time_zone_id field.

pub fn time_zone_id(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ()>

Writes the value of the time_zone_id field as a string.

Errors if the time_zone_id field is empty.

pub fn try_set_meta_zone_id(
&mut self,
id: &str
) -> DiplomatResult<(), ICU4XError>

Sets the meta_zone_id field from a string.

Errors if the string is not a valid BCP-47 meta zone ID.

pub fn clear_meta_zone_id(&mut self)

Clears the meta_zone_id field.

pub fn meta_zone_id(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ()>

Writes the value of the meta_zone_id field as a string.

Errors if the meta_zone_id field is empty.

pub fn try_set_zone_variant(
&mut self,
id: &str
) -> DiplomatResult<(), ICU4XError>

Sets the zone_variant field from a string.

Errors if the string is not a valid zone variant.

pub fn clear_zone_variant(&mut self)

Clears the zone_variant field.

pub fn zone_variant(
&self,
write: &mut DiplomatWriteable
) -> DiplomatResult<(), ()>

Writes the value of the zone_variant field as a string.

Errors if the zone_variant field is empty.

pub fn set_standard_time(&mut self)

Sets the zone_variant field to standard time.

pub fn set_daylight_time(&mut self)

Sets the zone_variant field to daylight time.

pub fn is_standard_time(&self) -> DiplomatResult<bool, ()>

Returns whether the zone_variant field is standard time.

Errors if the zone_variant field is empty.

pub fn is_daylight_time(&self) -> DiplomatResult<bool, ()>

Returns whether the zone_variant field is daylight time.

Errors if the zone_variant field is empty.

pub fn maybe_set_meta_zone(
&mut self,
local_datetime: &ICU4XIsoDateTime,
metazone_calculator: &ICU4XMetaZoneCalculator
)

Sets the meta zone based on the time zone and the local timestamp.