icu_capi::unstable::date::ffi

Struct IsoDate

#[repr(transparent)]
pub struct IsoDate(pub Date<Iso>);

An ICU4X Date object capable of containing a ISO-8601 date

Tuple Fields
0: Date<Iso>
Implementations

impl<'transparent_convert_outer> IsoDate

pub(crate) fn transparent_convert( from: &'transparent_convert_outer Date<Iso>, ) -> &'transparent_convert_outer Self

impl IsoDate

pub fn create( year: i32, month: u8, day: u8, ) -> Result<Box<IsoDate>, CalendarError>

Creates a new IsoDate from the specified date.

pub fn from_rata_die(rd: i64) -> Box<IsoDate>

Creates a new IsoDate from the given Rata Die

pub fn from_string(v: &DiplomatStr) -> Result<Box<IsoDate>, Rfc9557ParseError>

Creates a new IsoDate from an IXDTF string.

pub fn to_calendar(&self, calendar: &Calendar) -> Box<Date>

Convert this date to one in a different calendar

pub fn to_any(&self) -> Box<Date>

pub fn to_rata_die(&self) -> i64

Returns this date’s Rata Die

pub fn day_of_year(&self) -> u16

Returns the 1-indexed day in the year for this date

pub fn day_of_month(&self) -> u8

Returns the 1-indexed day in the month for this date

pub fn day_of_week(&self) -> Weekday

Returns the day in the week for this day

pub fn week_of_year(&self) -> IsoWeekOfYear

Returns the week number in this year, using week data

pub fn month(&self) -> u8

Returns 1-indexed number of the month of this date in its year

pub fn year(&self) -> i32

Returns the year number in the current era for this date

For calendars without an era, returns the extended year

pub fn is_in_leap_year(&self) -> bool

Returns if the year is a leap year for this date

pub fn months_in_year(&self) -> u8

Returns the number of months in the year represented by this date

pub fn days_in_month(&self) -> u8

Returns the number of days in the month represented by this date

pub fn days_in_year(&self) -> u16

Returns the number of days in the year represented by this date