Module icu_capi::properties_sets::ffi

Structs

ICU4XCodePointSetData

An ICU4X Unicode Set Property object, capable of querying whether a code point is contained in a set based on a Unicode property.

Functions

ICU4XCodePointSetData_contains ğŸ”’

ICU4XCodePointSetData_contains32 ğŸ”’

ICU4XCodePointSetData_destroy ğŸ”’

ICU4XCodePointSetData_iter_ranges ğŸ”’

ICU4XCodePointSetData_iter_ranges_complemented ğŸ”’

ICU4XCodePointSetData_load_alnum ğŸ”’

ICU4XCodePointSetData_load_alphabetic ğŸ”’

ICU4XCodePointSetData_load_ascii_hex_digit ğŸ”’

ICU4XCodePointSetData_load_bidi_control ğŸ”’

ICU4XCodePointSetData_load_bidi_mirrored ğŸ”’

ICU4XCodePointSetData_load_blank ğŸ”’

ICU4XCodePointSetData_load_case_ignorable ğŸ”’

ICU4XCodePointSetData_load_case_sensitive ğŸ”’

ICU4XCodePointSetData_load_cased ğŸ”’

ICU4XCodePointSetData_load_changes_when_casefolded ğŸ”’

ICU4XCodePointSetData_load_changes_when_casemapped ğŸ”’

ICU4XCodePointSetData_load_changes_when_lowercased ğŸ”’

ICU4XCodePointSetData_load_changes_when_nfkc_casefolded ğŸ”’

ICU4XCodePointSetData_load_changes_when_titlecased ğŸ”’

ICU4XCodePointSetData_load_changes_when_uppercased ğŸ”’

ICU4XCodePointSetData_load_dash ğŸ”’

ICU4XCodePointSetData_load_default_ignorable_code_point ğŸ”’

ICU4XCodePointSetData_load_deprecated ğŸ”’

ICU4XCodePointSetData_load_diacritic ğŸ”’

ICU4XCodePointSetData_load_emoji ğŸ”’

ICU4XCodePointSetData_load_emoji_component ğŸ”’

ICU4XCodePointSetData_load_emoji_modifier ğŸ”’

ICU4XCodePointSetData_load_emoji_modifier_base ğŸ”’

ICU4XCodePointSetData_load_emoji_presentation ğŸ”’

ICU4XCodePointSetData_load_extended_pictographic ğŸ”’

ICU4XCodePointSetData_load_extender ğŸ”’

ICU4XCodePointSetData_load_for_ecma262 ğŸ”’

ICU4XCodePointSetData_load_for_general_category_group ğŸ”’

ICU4XCodePointSetData_load_full_composition_exclusion ğŸ”’

ICU4XCodePointSetData_load_graph ğŸ”’

ICU4XCodePointSetData_load_grapheme_base ğŸ”’

ICU4XCodePointSetData_load_grapheme_extend ğŸ”’

ICU4XCodePointSetData_load_grapheme_link ğŸ”’

ICU4XCodePointSetData_load_hex_digit ğŸ”’

ICU4XCodePointSetData_load_hyphen ğŸ”’

ICU4XCodePointSetData_load_id_continue ğŸ”’

ICU4XCodePointSetData_load_id_start ğŸ”’

ICU4XCodePointSetData_load_ideographic ğŸ”’

ICU4XCodePointSetData_load_ids_binary_operator ğŸ”’

ICU4XCodePointSetData_load_ids_trinary_operator ğŸ”’

ICU4XCodePointSetData_load_join_control ğŸ”’

ICU4XCodePointSetData_load_logical_order_exception ğŸ”’

ICU4XCodePointSetData_load_lowercase ğŸ”’

ICU4XCodePointSetData_load_math ğŸ”’

ICU4XCodePointSetData_load_nfc_inert ğŸ”’

ICU4XCodePointSetData_load_nfd_inert ğŸ”’

ICU4XCodePointSetData_load_nfkc_inert ğŸ”’

ICU4XCodePointSetData_load_nfkd_inert ğŸ”’

ICU4XCodePointSetData_load_noncharacter_code_point ğŸ”’

ICU4XCodePointSetData_load_pattern_syntax ğŸ”’

ICU4XCodePointSetData_load_pattern_white_space ğŸ”’

ICU4XCodePointSetData_load_prepended_concatenation_mark ğŸ”’

ICU4XCodePointSetData_load_print ğŸ”’

ICU4XCodePointSetData_load_quotation_mark ğŸ”’

ICU4XCodePointSetData_load_radical ğŸ”’

ICU4XCodePointSetData_load_regional_indicator ğŸ”’

ICU4XCodePointSetData_load_segment_starter ğŸ”’

ICU4XCodePointSetData_load_sentence_terminal ğŸ”’

ICU4XCodePointSetData_load_soft_dotted ğŸ”’

ICU4XCodePointSetData_load_terminal_punctuation ğŸ”’

ICU4XCodePointSetData_load_unified_ideograph ğŸ”’

ICU4XCodePointSetData_load_uppercase ğŸ”’

ICU4XCodePointSetData_load_variation_selector ğŸ”’

ICU4XCodePointSetData_load_white_space ğŸ”’

ICU4XCodePointSetData_load_xdigit ğŸ”’

ICU4XCodePointSetData_load_xid_continue ğŸ”’

ICU4XCodePointSetData_load_xid_start ğŸ”’