Struct icu_capi::timezone::ffi::ICU4XCustomTimeZone

pub struct ICU4XCustomTimeZone(pub CustomTimeZone);

Tuple Fields
0: CustomTimeZone
Implementations

impl ICU4XCustomTimeZone

pub fn create_from_string( s: &DiplomatStr, ) -> Result<Box<ICU4XCustomTimeZone>, ICU4XError>

Creates a time zone from an offset string.

pub fn create_empty() -> Box<ICU4XCustomTimeZone>

Creates a time zone with no information.

pub fn create_utc() -> Box<ICU4XCustomTimeZone>

Creates a time zone for UTC.

pub fn try_set_gmt_offset_seconds( &mut self, offset_seconds: i32, ) -> Result<(), ICU4XError>

Sets the gmt_offset field from offset seconds.

Errors if the offset seconds are out of range.

pub fn clear_gmt_offset(&mut self)

Clears the gmt_offset field.

pub fn gmt_offset_seconds(&self) -> Result<i32, ICU4XError>

Returns the value of the gmt_offset field as offset seconds.

Errors if the gmt_offset field is empty.

pub fn is_gmt_offset_positive(&self) -> Result<bool, ICU4XError>

Returns whether the gmt_offset field is positive.

Errors if the gmt_offset field is empty.

pub fn is_gmt_offset_zero(&self) -> Result<bool, ICU4XError>

Returns whether the gmt_offset field is zero.

Errors if the gmt_offset field is empty (which is not the same as zero).

pub fn gmt_offset_has_minutes(&self) -> Result<bool, ICU4XError>

Returns whether the gmt_offset field has nonzero minutes.

Errors if the gmt_offset field is empty.

pub fn gmt_offset_has_seconds(&self) -> Result<bool, ICU4XError>

Returns whether the gmt_offset field has nonzero seconds.

Errors if the gmt_offset field is empty.

pub fn try_set_time_zone_id( &mut self, id: &DiplomatStr, ) -> Result<(), ICU4XError>

Sets the time_zone_id field from a BCP-47 string.

Errors if the string is not a valid BCP-47 time zone ID.

pub fn try_set_iana_time_zone_id( &mut self, mapper: &ICU4XIanaToBcp47Mapper, id: &DiplomatStr, ) -> Result<(), ICU4XError>

Sets the time_zone_id field from an IANA string by looking up the corresponding BCP-47 string.

Errors if the string is not a valid BCP-47 time zone ID.

pub fn try_set_iana_time_zone_id_2( &mut self, mapper: &ICU4XTimeZoneIdMapper, id: &DiplomatStr, ) -> Result<(), ICU4XError>

Sets the time_zone_id field from an IANA string by looking up the corresponding BCP-47 string.

Errors if the string is not a valid BCP-47 time zone ID.

pub fn clear_time_zone_id(&mut self)

Clears the time_zone_id field.

pub fn time_zone_id( &self, write: &mut DiplomatWriteable, ) -> Result<(), ICU4XError>

Writes the value of the time_zone_id field as a string.

Errors if the time_zone_id field is empty.

pub fn try_set_metazone_id( &mut self, id: &DiplomatStr, ) -> Result<(), ICU4XError>

Sets the metazone_id field from a string.

Errors if the string is not a valid BCP-47 metazone ID.

pub fn clear_metazone_id(&mut self)

Clears the metazone_id field.

pub fn metazone_id( &self, write: &mut DiplomatWriteable, ) -> Result<(), ICU4XError>

Writes the value of the metazone_id field as a string.

Errors if the metazone_id field is empty.

pub fn try_set_zone_variant( &mut self, id: &DiplomatStr, ) -> Result<(), ICU4XError>

Sets the zone_variant field from a string.

Errors if the string is not a valid zone variant.

pub fn clear_zone_variant(&mut self)

Clears the zone_variant field.

pub fn zone_variant( &self, write: &mut DiplomatWriteable, ) -> Result<(), ICU4XError>

Writes the value of the zone_variant field as a string.

Errors if the zone_variant field is empty.

pub fn set_standard_time(&mut self)

Sets the zone_variant field to “standard” time, which may or may not correspond to a display name with “Standard” in its name.

pub fn set_daylight_time(&mut self)

Sets the zone_variant field to “daylight” time, which may or may not correspond to a display name with “Daylight” in its name.

pub fn is_standard_time(&self) -> Result<bool, ICU4XError>

Returns whether the zone_variant field is standard time.

Errors if the zone_variant field is empty.

pub fn is_daylight_time(&self) -> Result<bool, ICU4XError>

Returns whether the zone_variant field is daylight time.

Errors if the zone_variant field is empty.

pub fn maybe_calculate_metazone( &mut self, metazone_calculator: &ICU4XMetazoneCalculator, local_datetime: &ICU4XIsoDateTime, )

Sets the metazone based on the time zone and the local timestamp.